+++ scripts/triggers-end-to-end-tests/v2/index.js
@@ -23,9 +23,12 @@ const AUTH_BLOCKING_CREATE_V2_LOG =
 const AUTH_BLOCKING_SIGN_IN_V2_LOG =
   "========== AUTH BLOCKING SIGN IN V2 FUNCTION METADATA ==========";
 const RTDB_LOG = "========== RTDB V2 FUNCTION ==========";
+const FIRESTORE_LOG = "========== FIRESTORE V2 FUNCTION ==========";
 
 const PUBSUB_TOPIC = "test-topic";
+
 const START_DOCUMENT_NAME = "test/start";
+const END_DOCUMENT_NAME = "test/done";
 
 admin.initializeApp();
 
@@ -136,3 +139,26 @@ exports.rtdbv2reaction = functionsV2.database.onValueWritten(START_DOCUMENT_NAME
   console.log(RTDB_LOG);
   return;
 });
+
+exports.firestorev2reaction = functionsV2.firestore.onDocumentWritten(
+  START_DOCUMENT_NAME,
+  async (event) => {
+    console.log(FIRESTORE_LOG);
+    /*
+     * Write back a completion timestamp to the firestore emulator. The test
+     * driver program checks for this by querying the firestore emulator
+     * directly.
+     */
+    const ref = admin.firestore().doc(END_DOCUMENT_NAME + "_from_firestore");
+    await ref.set({ done: new Date().toISOString() });
+
+    /*
+     * Write a completion marker to the firestore emulator. This exercise
+     * cross-emulator communication.
+     */
+    const dbref = admin.database().ref(END_DOCUMENT_NAME + "_from_firestore");
+    await dbref.set({ done: new Date().toISOString() });
+
+    return true;
+  }
+);