+++ CHANGELOG.md
@@ -2,3 +2,4 @@
 - Upgrade Firestore emulator to 1.17.4
 - Web Frameworks will no longer try to deploy unsupported versions of NodeJS to Cloud Functions (#5733)
 - Adding exerimental support for deploying Flutter Web applications to Firebase Hosting (#5332)
+- Adds function emulator support for v2 firestore triggers (#5685).